revises on the sh scripts

+++ shortRange.inhomo/template/run.ball.init.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-source parameters.sh
+source p.collision.sh
 
 function set_parameter () {
     file=$1
@@ -9,6 +9,7 @@ function set_parameter () {
     sed -e "/IndexType thermoFeq/s/=.*/= $ball_thermoFeq;/g"|\
     sed -e "/ScalorType rcut/s/=.*/= $ball_rcut;/g" |\
     sed -e "/ScalorType nlistExten/s/=.*/= $ball_nlistExten;/g" |\
+    sed -e "/ScalorType nlistSizeFactor/s/=.*/= $ball_nlistSizeFactor;/g" |\
     sed -e "/ScalorType refT/s/=.*/= $ball_refT;/g" |\
     sed -e "/ScalorType tauT/s/=.*/= $ball_tauT;/g" |\
     sed -e "/ScalorType dt/s/=.*/= $ball_dt;/g" |\
@@ -20,10 +21,17 @@ function set_parameter () {
 
 set_parameter tools/simulation.uniform.cutoff/app/lj.nhc.cu
 make -j4 -C tools/simulation.uniform.cutoff/ &> make.log
+
 rm -f conf.gro
-make -j4 -C tools.collision/gen.conf/ &> make.log
-tools.collision/gen.conf/fcc.ball
-mv confout.gro conf.gro
+if test -f $ball_init_conf; then
+    echo "# the init conf exists"
+    cp $ball_init_conf ./conf.gro
+else
+    echo "# the init conf does not exist, generate one"
+    make -j4 -C tools.collision/gen.conf/ &> make.log
+    tools.collision/gen.conf/fcc.ball
+    mv confout.gro conf.gro
+fi
 workdir=`pwd`
 command="$workdir/tools/simulation.uniform.cutoff/lj.nhc conf.gro $ball_nstep $device"
 echo "command is: $command > gpu-md.out 2> gpu-md.perf"
